requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.2
django-coverage-plugin==1.3.1

# Used for converting .srt to .md (might not be required)
recommonmark==0.4.0

Sphinx==1.4.5
django-extensions==1.7.4
Werkzeug==0.11.11
django-test-plus==1.0.15
factory_boy==2.7.0

django-debug-toolbar==1.7

# improved REPL
ipdb==0.10.1

# Tests
pytest-django==2.9.1
pytest-sugar==0.7.1
pytest-cov==2.4.0
pytest-watch==4.1.0
pytest-testmon==0.9.4
